neatogen call_tri2: take 'm' as a 'size_t'
This pays down some debt from previous changes. Squashes 2
-Wsign-conversion warnings.

+++ lib/sfdpgen/post_process.c
@@ -777,7 +777,7 @@ TriangleSmoother TriangleSmoother_new(SparseMatrix A, int dim, double *x,
     if (use_triangularization) {
       B = call_tri((int)m, x);
     } else {
-      B = call_tri2((int)m, dim, x);
+      B = call_tri2(m, dim, x);
     }
   } else {
     B = SparseMatrix_copy(A);
